fix(schema): add missing admin checks (#3173)
* fix(schema): add missing admin checks

* update

+++ ee/tabby-schema/src/schema/mod.rs
@@ -491,6 +491,7 @@ impl Query {
         first: Option<i32>,
         last: Option<i32>,
     ) -> Result<Connection<Integration>> {
+        check_admin(ctx).await?;
         query_async(
             after,
             before,
@@ -516,6 +517,7 @@ impl Query {
         first: Option<i32>,
         last: Option<i32>,
     ) -> Result<Connection<ProvidedRepository>> {
+        check_admin(ctx).await?;
         query_async(
             after,
             before,
@@ -592,6 +594,7 @@ impl Query {
         first: Option<i32>,
         last: Option<i32>,
     ) -> Result<Connection<CustomWebDocument>> {
+        check_admin(ctx).await?;
         query_async(
             after,
             before,
@@ -615,6 +618,7 @@ impl Query {
         last: Option<i32>,
         is_active: Option<bool>,
     ) -> Result<Connection<PresetWebDocument>> {
+        check_admin(ctx).await?;
         query_async(
             after,
             before,
@@ -939,6 +943,7 @@ impl Mutation {
     }
 
     async fn create_integration(ctx: &Context, input: CreateIntegrationInput) -> Result<ID> {
+        check_admin(ctx).await?;
         input.validate()?;
         let id = ctx
             .locator
@@ -954,6 +959,7 @@ impl Mutation {
     }
 
     async fn update_integration(ctx: &Context, input: UpdateIntegrationInput) -> Result<bool> {
+        check_admin(ctx).await?;
         input.validate()?;
         ctx.locator
             .integration()
@@ -969,6 +975,7 @@ impl Mutation {
     }
 
     async fn delete_integration(ctx: &Context, id: ID, kind: IntegrationKind) -> Result<bool> {
+        check_admin(ctx).await?;
         ctx.locator
             .integration()
             .delete_integration(id, kind)
@@ -981,6 +988,7 @@ impl Mutation {
         id: ID,
         active: bool,
     ) -> Result<bool> {
+        check_admin(ctx).await?;
         ctx.locator
             .repository()
             .third_party()
@@ -1035,6 +1043,7 @@ impl Mutation {
     }
 
     async fn create_custom_document(ctx: &Context, input: CreateCustomDocumentInput) -> Result<ID> {
+        check_admin(ctx).await?;
         input.validate()?;
         let id = ctx
             .locator
@@ -1045,6 +1054,7 @@ impl Mutation {
     }
 
     async fn delete_custom_document(ctx: &Context, id: ID) -> Result<bool> {
+        check_admin(ctx).await?;
         ctx.locator
             .web_documents()
             .delete_custom_web_document(id)
@@ -1056,6 +1066,7 @@ impl Mutation {
         ctx: &Context,
         input: SetPresetDocumentActiveInput,
     ) -> Result<bool> {
+        check_admin(ctx).await?;
         input.validate()?;
         ctx.locator
             .web_documents()
